requirements/prod.txt
python-decouple==3.8
SQLAlchemy==2.0.32
vkbottle==4.3.12
